### 2017-09-07 #397 @arfon
👋 @openjournals/ropensci - would this submission be a good candidate for rOpenSci?
