### 2017-09-10 #397 @leeper
This looks like a good candidate for rOpenSci. If it doesn't go that route, I can be editor.
