### 2017-09-11 #397 @karthik
@leeper This is a pass from ropensci, so please continue with a JOSS review.
