### 2017-09-21 #397 @leeper
've confirmed that this submission has:

- [X]  An OSI approved license (GPL)
- [X] A description of the software and its research applications
- [X] Is in scope for this journal
