### 2017-09-21 #408 @whedon
[REVIEW]: UKCensusAPI: python and R interfaces to the nomisweb UK census data API
**Submitting author:** @virgesmith (<a href="http://orcid.org/0000-0002-9951-6642">Andrew Smith</a>)
**Repository:** <a href="https://github.com/virgesmith/UKCensusAPI" target ="_blank">https://github.com/virgesmith/UKCensusAPI</a>
**Version:** 0.1
