### 2017-09-22 d4f1743
Try to fix travis R build error
