### 2017-09-22 1ececce
Try forcing pip3 install
