+++ .travis.yml
@@ -1,6 +1,6 @@
 
 cache: packages
-sudo: true
+sudo: false
 
 matrix:
   include:
@@ -12,7 +12,7 @@ matrix:
       before_install:
         - pip3 install numpy pandas
         - ./setup.py install
-      dist: trusty
+      dist: precise
 
     - language: r
       r: release
