+++ .travis.yml
@@ -6,9 +6,14 @@ matrix:
   include:
     - language: r
       r: oldrel
+      # install the python package first
+      addons:
+        apt:
+          packages:
+          - "python3"
+          - "python3-pip"
       env:
         - RETICULATE_PYTHON=$(which python3)
-      # install the python package first
       before_install:
         - pip3 install numpy pandas
         - ./setup.py install
