+++ .travis.yml
@@ -26,11 +26,13 @@ matrix:
           packages:
           - "python3"
           - "python3-pip"
+          - "python3-numpy"
+          - "python3-pandas"
       env:
         - RETICULATE_PYTHON=$(which python3)
       # install the python package first
       before_install:
-        - pip3 install .
+       # - pip3 install .
        # -r requirements.txt
         - ./setup.py install
       dist: trusty
