+++ .travis.yml
@@ -24,7 +24,7 @@ matrix:
       addons:
         apt:
           packages:
-          - "python3"
+          - "python3.5"
           - "python3-pip"
           - "python3-numpy"
           - "python3-pandas"
