+++ .travis.yml
@@ -19,23 +19,23 @@ matrix:
     #     - ./setup.py install
     #   dist: trusty
 
-    - language: r
-      r: release
-      addons:
-        apt:
-          packages:
-          - "python3.5"
-          - "python3-pip"
-          - "python3-numpy"
-          - "python3-pandas"
-      env:
-        - RETICULATE_PYTHON=$(which python3)
-      # install the python package first
-      before_install:
-       # - pip3 install .
-       # -r requirements.txt
-        - ./setup.py install
-      dist: trusty
+    # - language: r
+    #   r: release
+    #   addons:
+    #     apt:
+    #       packages:
+    #       - "python3"
+    #       - "python3-pip"
+    #       - "python3-numpy"
+    #       - "python3-pandas"
+    #   env:
+    #     - RETICULATE_PYTHON=$(which python3)
+    #   # install the python package first
+    #   before_install:
+    #    # - pip3 install .
+    #    # -r requirements.txt
+    #     - ./setup.py install
+    #   dist: trusty
 
     # - language: r
     #   r: devel
@@ -47,26 +47,26 @@ matrix:
     #     - ./setup.py install
     #   dist: trusty
 
-    # - language: python
-    #   python: 3.4
-    #   install:
-    #     - pip install -r requirements.txt
-    #   script:
-    #     - python setup.py test
+    - language: python
+      python: 3.4
+      install:
+        - pip install -r requirements.txt
+      script:
+        - python setup.py test
 
-    # - language: python
-    #   python: 3.5
-    #   install:
-    #     - pip install -r requirements.txt
-    #   script:
-    #     - python setup.py test
+    - language: python
+      python: 3.5
+      install:
+        - pip install -r requirements.txt
+      script:
+        - python setup.py test
 
-    # - language: python
-    #   python: 3.6
-    #   install:
-    #     - pip install -r requirements.txt
-    #   script:
-    #     - python setup.py test
+    - language: python
+      python: 3.6
+      install:
+        - pip install -r requirements.txt
+      script:
+        - python setup.py test
 
 warnings_are_errors: false
 
