+++ .travis.yml
@@ -44,7 +44,7 @@ matrix:
       # install the python package first
       before_install:
         - pyenv global system 3.5
-        - pip3 install numpy pandas
+        - pip3 install setuptools numpy pandas
         - ./setup.py install
       dist: trusty
 
