### 2017-09-22 ff14316
revert r build back to old working config and allow failures until travis environment is fixed
