### 2017-09-22 2ccd875
see if this fixes: https://www.reddit.com/r/learnpython/comments/6zbywe/the_pip3_command_exists_in_these_python_versions/
