### 2017-09-22 7523bca
promote non-dev r builds to failable. remove debug print
