+++ .travis.yml
@@ -59,8 +59,6 @@ matrix:
         - python setup.py test
 
   allow_failures:
-    - r: oldrel
-    - r: release
     - r: devel
 
 warnings_are_errors: false
