### 2017-10-03 virgesmith/UKCensusAPI#14 @jrnold
Create a release tag
When submitting to JOSS the release version should be [tagged in GitHub](https://help.github.com/articles/creating-releases/) and match the version number in setup.py and also DESCRIPTION.

This is part of the JOSS review process: openjournals/joss-reviews#408.
