### 2017-10-03 virgesmith/UKCensusAPI#15 @jrnold
Split the R and Python Packages
Having the R and Python package share the same directory is confusing and could possibly lead to unforeseen release issues.

- Split into two repositories. This is clean, and usually people expect a packages to be in their own repository. Repositories are cheap (free)
- If they must be kept in the same repository, put the R and python packages in different directories.

This is part of the JOSS review openjournals/joss-reviews#408
