### 2017-10-03 virgesmith/UKCensusAPI#16 @jrnold
Suggested edits for README
- In the first sentence or paragraph describe the what this is accessing. The term nomisweb is used (and only linked later). Presumably, someone looking for this repository knows what that it, but since nomisweb != UK Census, it was confusing.
- Instead of `python3`, use a particular python version Python (>= 3.3). If Python 2 is not supported (and there is no reason it should be anymore, it may be worth explicitly stating that).

This is part of the JOSS review openjournals/joss-reviews#408.
