### 2017-10-03 virgesmith/UKCensusAPI#15 @jrnold
Combining the tests for both R and python in the same directory: https://github.com/virgesmith/UKCensusAPI/tree/master/tests is not ideal since under some configurations R will install the unrelated python tests.
