### 2017-10-03 virgesmith/UKCensusAPI#16 @jrnold
In addition, there should be a snippets with the name of the R package and python module to load. This is especially the case since the R package name and python package names are not the same as the repo name (`"ukcensusapi" != "UKCensusAPI"`):
```python
import ukcensusapi
```
```R
library("ukcensusapi")
```
