### 2017-10-04 8cd3651
address review comments re: documentation #16
