### 2017-10-05 a901f33
sorting out python code doc #10
