+++ ukcensusapi/__init__.py
@@ -1,3 +1,2 @@
-# TODO work out how to best deal with versioning
 __version__ = "1.0.0"
 
