### 2017-10-05 virgesmith/UKCensusAPI#14 @virgesmith
Since tags can't be modified I shall leave this until last
