### 2017-10-05 virgesmith/UKCensusAPI#15 @virgesmith
I'm loath to split the packages, as the idea of having a single package that works in both R and python was a key driver to creating this package. If split they will inevitably diverge, and the R package will have a dependency on a separate non-R package which will complicate installation.

I appreciate some compromises have to be made due to package structure restrictions - the scripts need be be under the `inst` directory or R won't install them, and the test directory (as you point out) contains both R and python scripts, but on balance I don't see this as a major issue - unless there's something I haven't thought of?
