### 2017-10-09 793e664
add description fields, correct licence
