+++ DESCRIPTION
@@ -1,12 +1,17 @@
 Package: UKCensusAPI
-Title: What the Package Does (one line, title case)
-Version: 0.0.0.9000
-Authors@R: person("First", "Last", email = "first.last@example.com", role = c("aut", "cre"))
-Description: What the package does (one paragraph).
+Title: Automated Query and Download of UK Census Data  
+Version: 1.0.0
+Authors@R: person("Andrew", "Smith", email = "a.p.smith@leeds.ac.uk", role = c("aut", "cre"))
+Description: This package provides an R interface to the www.nomisweb.co.uk census data API, enabling:
+    - querying table metadata
+    - autogenerating customised python and R query code for future use
+    - automated cached data downloads
+    - easily modifying the geography of existing queries
+    - adding descriptive information to tables (from metadata)
 Depends: R (>= 3.3.3)
 Imports:
   reticulate,
   testthat,
-License: GPL-3
+License: MIT + file LICENCE
 Encoding: UTF-8
 RoxygenNote: 6.0.1
