### 2017-10-09 843d384
improvements to geo codes #6
