### 2017-10-09 300884f
added 2001 OA support and fixed interactive query code #6
