### 2017-10-30 #408 @virgesmith
@leeper: I've addressed issues raised so far by @jrnold. Two are still open: #15 (split into 2 packages), because in my opinion this isn't a good idea (I've explained my position), and #14 (release tag) as I wanted to leave this until all other issues were resolved. I'm really waiting for @jrnold to respond on these (and any further issues he wants to raise)
