### 2017-11-10 #408 @virgesmith
@leeper, @jrnold I'm keen to progress this but its all gone rather quiet...
