### 2017-11-10 #408 @leeper
@virgesmith I've just taken another look at this. I would like to see some small revisions before accepting this. Namely, it would be good for all of the R documentation pages to include small examples of code usage. It's fine to include all of these in `\dontrun{}` brackets. Similarly, can the main package [documentation page](https://github.com/virgesmith/UKCensusAPI/blob/master/R/UKCensusAPI.R) include some of the general introductory content from the README? READMEs are not easily accessible from within R so having this within-package documentation is pretty important from a usability perspective.

Also please do generate a github release (per https://github.com/virgesmith/UKCensusAPI/issues/14) and initiate a Zenodo (or other) archive to obtain a DOI for that release.
