### 2017-11-13 57e42c6
added examples as per #19
