+++ R/Nomisweb.R
@@ -3,6 +3,8 @@
 #'
 #' This function calls an interactive script where the user selects a table, a geography, and selects fields, optionally filtering by value.
 #' This script will not run in RStudio due to the way it handles standard input. Please run in a standalone R session (or call the python script directly)
+#' @examples
+#' \dontrun{queryMetadata()}
 #' @export
 queryMetadata = function() {
   # first check we are not running in RStudio (in which can we cannot run interatively, since RStudio redirects stdin from /dev/null)
@@ -16,9 +18,14 @@ queryMetadata = function() {
 
 #' Geographic code lookup
 #'
-#' This function returns the nomisweb code for a particular geographic area
+#' This function returns the nomisweb code for a particular geographic area type
 #' @param api an instance of the API (returned by instance())
 #' @param geoCodeString the string representation of a geography, e.g MSOA11
+#' @return code - an integer
+#' @examples
+#' library(UKCensusAPI)
+#' censusapi = instance("/tmp/UKCensusAPI")
+#' geoCodeLookup(censusapi, "MSOA11")
 #' @export
 geoCodeLookup = function(api, geoCodeString) {
   return(as.integer(api$GeoCodeLookup[geoCodeString]))
@@ -27,9 +34,16 @@ geoCodeLookup = function(api, geoCodeString) {
 #' getMetadata()
 #' Fetch the metadata for a census table
 #'
-#' @export
 #' @param api the census provider api
 #' @param tableName the name of the census table
+#' @return metadata a list data structure containing the table metadata
+#' @examples
+#' \dontrun{
+#' library(UKCensusAPI)
+#' censusapi = instance("/tmp/UKCensusAPI")
+#' getMetadata(censusapi, "KS001")
+#' }
+#' @export
 getMetadata = function(api, tableName) {
   return(api$get_metadata(tableName))
 }
@@ -43,6 +57,24 @@ getMetadata = function(api, tableName) {
 #' @param tableName name of census table (e.g KS401EW)
 #' @param internalName internal name of census table (e.g NM_618_1)
 #' @param query query parameters
+#' @return a data.frame contraing the downloaded data
+#' @examples
+#' \dontrun{
+#' library(UKCensusAPI)
+#' censusapi = instance("/tmp/UKCensusAPI")
+#' table = "KS102EW"
+#' meta=getMetadata(censusapi, table)
+#' # queryParams can be autogenerated using the interactive query functionality
+#' queryParams = list(
+#'   geography = "1249902593...1249902596,1249934513...1249934514",
+#'   MEASURES = "20100",
+#'   select = "GEOGRAPHY_CODE,OBS_VALUE",
+#'   CELL = "0",
+#'   RURAL_URBAN = "0",
+#'   date = "latest"
+#' )
+#' getData(censusapi, table, meta$nomis_table, queryParams)
+#' }
 #' @export
 getData = function(api, tableName, internalName, query) {
   # returned value is filename (or error) to avoid data frame compatibility issues
@@ -51,11 +83,15 @@ getData = function(api, tableName, internalName, query) {
   return(read.csv(filename, sep="\t", stringsAsFactors = FALSE))
 }
 
-#' Map local authority names to nomisweb codes (python)
+#' Map local authority names to nomisweb codes
 #'
 #' @param api an instance of the UKCensusData API.
 #' @param laNames a string vector of local authority names.
 #' @return an integer vector of nomisweb local authority codes
+#' @examples
+#' library(UKCensusAPI)
+#' censusapi = instance("/tmp/UKCensusAPI")
[…]
