+++ man/contextify.Rd
@@ -21,3 +21,22 @@ the table containing a new column with the contextual data
 \description{
 Append table with a contextual column.
 }
+\examples{
+\dontrun{
+library("UKCensusAPI")
+cacheDir = "/tmp/UKCensusAPI/"
+censusapi = UKCensusAPI::instance(cacheDir)
+table = "KS401EW"
+table_internal = "NM_618_1"
+queryParams = list(
+  date = "latest",
+  CELL = "7...13", # dwelling type
+  select = "GEOGRAPHY_CODE,CELL,OBS_VALUE",
+  MEASURES = "20100",
+  geography = "1245710558...1245710560",
+  RURAL_URBAN = "0"
+)
+KS401EW = UKCensusAPI::getData(censusapi, table, table_internal, queryParams)
+annotated = contextify(censusapi, table, "CELL", KS401EW)
+}
+}
