+++ DESCRIPTION
@@ -2,7 +2,7 @@ Package: UKCensusAPI
 Title: Automated Query and Download of UK Census Data  
 Version: 1.0.0
 Authors@R: person("Andrew", "Smith", email = "a.p.smith@leeds.ac.uk", role = c("aut", "cre"))
-Description: This package provides an R interface to the www.nomisweb.co.uk census data API, enabling:
+Description: This package provides an R interface to the www.nomisweb.co.uk census data API. It enables:
     - querying table metadata
     - autogenerating customised python and R query code for future use
     - automated cached data downloads
