### 2017-11-13 virgesmith/UKCensusAPI#14 @virgesmith
Created 1.0.0 release
