### 2017-11-14 #408 @virgesmith
@leeper have done the documentation changes and the release as requested.
