### 2017-11-21 #408 @arfon
@virgesmith - your paper is now accepted into JOSS and your paper DOI is https://doi.org/10.21105/joss.00408 ⚡️ 🚀 💥
