### 2017-11-21 #408 @arfon
@virgesmith - also, please merge these changes that I had to make to complete your submission: https://github.com/virgesmith/UKCensusAPI/pull/20
